follow up changing User ID from Integer to Long

+++ RR-Common/src/main/java/com/scor/rr/repository/UserRrRepository.java
@@ -4,7 +4,7 @@ import com.scor.rr.domain.UserRrEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface UserRrRepository extends JpaRepository<UserRrEntity, Integer> {
+public interface UserRrRepository extends JpaRepository<UserRrEntity, Long> {
 
     @Query("select c from UserRrEntity c where c.userId= :userId")
     UserRrEntity findByUserName(int userId);
